some bugfixes

+++ lib/ZnapZend.pm
@@ -401,15 +401,17 @@ my $daemonize = sub {
     my $self = shift;
     my $pidFile = $self->pidfile // $self->defaultPidFile;
 
-    #make sure pid file is writable
+    if (-f $pidFile){
+        chomp(my $pid = slurp $pidFile);
+        #pid is not empty and is numeric
+        if ($pid && ($pid = int($pid)) && kill 0, $pid){
+            die "I Quit! Another copy of znapzend ($pid) seems to be running. See $pidFile\n";
+        }
+    }
+    #make sure pid file is writable before forking
     open my $fh, '>', $pidFile or die "ERROR: pid file '$pidFile' is not writable\n";
     close $fh;
 
-    chomp(my $pid = slurp $pidFile);
-    #pid is not empty and is numeric
-    if ($pid && ($pid = int($pid)) && kill 0, $pid){
-        die "I Quit! Another copy of znapzend ($pid) seems to be running. See $pidFile\n";
-    }
     defined (my $pid = fork) or die "Can't fork: $!";
 
     if ($pid){
